requirements.txt
Appium-Python-Client==2.11.1
assertpy
pytest
selenium
